refactor: BundleViewModel 파일 구조 재정리 및 주석 추가
- +Types: 데이터 구조체 분리 (BackgroundViewData, CarabinerViewData, KeyringInfo)
- +Fetch: +LoadData에서 이름 변경
- +Edit: +Sort 병합
- +Purchase: 구매 로직 분리
- +Helpers: 유틸리티 메서드 분리
- +Cache: +ReloadDecision + 이미지 캐시 로직 병합
- +Views: +CaptureScene + 이미지 뷰빌더 병합
- 각 파일에 역할 설명 주석 추가
- 아 힘들어
