+++ Keychy/Keychy/Presentation/Bundle/ViewModels/BundleViewModel+Fetch.swift
@@ -1,17 +1,27 @@
 //
-//  BundleViewModel+LoadData.swift
+//  BundleViewModel+Fetch.swift
 //  Keychy
 //
 //  Created by 김서현 on 1/12/26.
 //
 
+// MARK: - BundleViewModel+Fetch
+//
+// Firebase 데이터 읽기
+// - fetchAllBundles: 사용자 뭉치 로드
+// - fetchAllBackgrounds: 배경 로드
+// - fetchAllCarabiners: 카라비너 로드
+// - fetchKeyringInfo: 키링 정보 로드
+
 import FirebaseFirestore
 
 extension BundleViewModel {
-    //MARK: - Firebase에서 사용자의 모든 뭉치 로드
+
+    // MARK: - 사용자의 모든 뭉치 로드
+
     func fetchAllBundles(uid: String, completion: @escaping (Bool) -> Void) {
         isLoading = true
-        
+
         db.collection("KeyringBundle")
             .whereField("userId", isEqualTo: uid)
             .getDocuments { [weak self] snapshot, error in
@@ -19,46 +29,44 @@ extension BundleViewModel {
                     completion(false)
                     return
                 }
-                
+
                 defer { self.isLoading = false }
-                
+
                 if let error = error {
                     print("뭉치 로드 에러: \(error.localizedDescription)")
                     completion(false)
                     return
                 }
-                
+
                 guard let documents = snapshot?.documents else {
                     print("뭉치 문서가 없습니다.")
                     self.bundles = []
                     completion(true)
                     return
                 }
-                
+
                 let loadedBundles: [KeyringBundle] = documents.compactMap { doc in
                     KeyringBundle(documentId: doc.documentID, data: doc.data())
                 }
-                
-                // 뷰모델 번들에 저장 (정렬은 sortedBundles에서 처리)
+
                 self.bundles = loadedBundles
                 completion(true)
             }
     }
-    
-    // MARK: - 전체 배경 로드 + 소유 여부 주석 (dataManager 활용)
+
+    // MARK: - 전체 배경 로드 + 소유 여부 표시
+
     func fetchAllBackgrounds(completion: @escaping (Bool) -> Void) {
         isLoading = true
         Task {
-            // dataManager를 통해 캐싱된 데이터 활용
             await dataManager.fetchBackgroundsIfNeeded()
-            
-            // dataManager에서 이미 로드된 데이터 가져오기
-            let items = backgrounds // dataManager.backgrounds
+
+            let items = backgrounds
             let ownedIds = UserManager.shared.currentUser?.backgrounds ?? []
             let decorated = items.map { bg in
                 BackgroundViewData(background: bg, isOwned: ownedIds.contains(bg.id ?? ""))
             }
-            
+
             await MainActor.run {
                 self.backgroundViewData = decorated
                 self.isLoading = false
@@ -66,32 +74,31 @@ extension BundleViewModel {
             }
         }
     }
-    
-    // MARK: - 전체 카라비너 로드 + 소유 여부 주석 (dataManager 활용)
+
+    // MARK: - 전체 카라비너 로드 + 소유 여부 표시
+
     func fetchAllCarabiners(completion: @escaping (Bool) -> Void) {
         isLoading = true
         Task {
-            // dataManager를 통해 캐싱된 데이터 활용
             await dataManager.fetchCarabinersIfNeeded()
-            
-            // dataManager에서 이미 로드된 데이터 가져오기
-            let items = carabiners // dataManager.carabiners
+
+            let items = carabiners
             let ownedIds = UserManager.shared.currentUser?.carabiners ?? []
             let decorated = items.map { cb in
                 CarabinerViewData(carabiner: cb, isOwned: ownedIds.contains(cb.id ?? ""))
             }
-            
+
             await MainActor.run {
                 self.carabinerViewData = decorated
                 self.isLoading = false
                 completion(true)
             }
         }
-        // 카라비너는 기본 카라비너 자동 선택 됨
         selectedCarabiner = carabiners.first
     }
-    
-    /// Firestore에서 키링 정보를 가져옴
+
+    // MARK: - 단일 키링 정보 로드
+
     func fetchKeyringInfo(keyringId: String) async -> KeyringInfo? {
         do {
             let document = try await db.collection("Keyring").document(keyringId).getDocument()
